Added all the logic to the site.
Started out with adding all the functions of operations.
Then added all the buttons to their respective variable.
Added an EventListener using ForEach on all the numbers adds the number
to the "display of the number".
added a EventListener to the operators too that clears the screen and
adds the number that was displayed to a variable and the operator
to a variable as well.

The Equal sign puts the numbers and operator into a function that
determines the operator and calls the corresponding function and returns it.
It then gets displayed onto the "display"