Fixed two bugs.
First bug was that the the add fucntion couldn't handle float numbers well
it was becuase it used parseInt instead of parseFloat.

Second bug was that the calculator couldn't handle calculating
pressing = before entering all of the numbers or an operator. Fixed by
adding if statements in the eventlistener for the equal button.